+++ utils/utils.go
@@ -1,6 +1,65 @@
 package utils
 
-import "strconv"
+import (
+	"context"
+	"fmt"
+	"net/http"
+	"regexp"
+	"strconv"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+	gwruntime "github.com/grpc-ecosystem/grpc-gateway/runtime"
+	"github.com/spf13/viper"
+	loggermiddleware "github.com/meateam/api-gateway/logger"
+	"github.com/meateam/api-gateway/user"
+	fpb "github.com/meateam/file-service/proto/file"
+	ppb "github.com/meateam/permission-service/proto"
+	"github.com/sirupsen/logrus"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
+)
+
+type Router struct {
+	logger          *logrus.Logger
+}
+
+type User struct {
+	ID          string `json:"id"`
+	FirstName   string `json:"firstname"`
+	LastName    string `json:"lastname"`
+	Source      string `json:"source"`
+	Bucket      string `json:"bucket"`
+	DisplayName string `json:"displayName"`
+	CurrentUnit string `json:"currentUnit"`
+	Rank        string `json:"rank"`
+	Job         string `json:"job"`
+}
+
+
+const (
+	// OwnerRole is the owner role name when referred to as a permission.
+	OwnerRole = "OWNER"
+
+	// ConfigBucketPostfix is the name of the environment variable containing the postfix for the bucket.
+	ConfigBucketPostfix = "bucket_postfix"
+
+	// ContextUserKey is the context key used to get and set the user's data in the context.
+	ContextUserKey = "User"
+)
+
+func NewRouter(
+	logger *logrus.Logger,
+) *Router {
+	// If no logger is given, use a default logger.
+	if logger == nil {
+		logger = logrus.New()
+	}
+	
+	r := &Router{logger: logger}
+
+	return r
+}
 
 // StringToInt64 converts a string to int64, 0 is returned on failure
 func StringToInt64(s string) int64 {
@@ -9,4 +68,163 @@ func StringToInt64(s string) int64 {
 		n = 0
 	}
 	return n
+}
+
+// ExtractRequestUser gets a context.Context and extracts the user's details from c.
+func ExtractRequestUser(ctx context.Context) *User {
+	contextUser := ctx.Value(ContextUserKey)
+	if contextUser == nil {
+		return nil
+	}
+
+	var reqUser User
+	switch v := contextUser.(type) {
+	case User:
+		reqUser = v
+		reqUser.Bucket = normalizeCephBucketName(reqUser.ID)
+	default:
+		return nil
+	}
+
+	return &reqUser
+}
+
+
+// normalizeCephBucketName gets a bucket name and normalizes it
+// according to ceph s3's constraints.
+func normalizeCephBucketName(bucketName string) string {
+	postfix := viper.GetString(ConfigBucketPostfix)
+	lowerCaseBucketName := strings.ToLower(bucketName + postfix)
+
+	// Make a Regex for catching only letters and numbers.
+	reg := regexp.MustCompile("[^a-zA-Z0-9]+")
+	return reg.ReplaceAllString(lowerCaseBucketName, "-")
+}
+
+
+// HandleUserFilePermission gets the id of the requested file, and the required role.
+// Returns the user role as a string, and the permission if the user is permitted
+// to operate on the file, and `"", nil` if not.
+func HandleUserFilePermission(
+	fileClient fpb.FileServiceClient, 
+	permissionClient ppb.PermissionClient,
+	c *gin.Context,
+	fileID string,
+	role ppb.Role) (string, *ppb.PermissionObject) {
+	r := NewRouter(&logrus.Logger{})
+	reqUser := user.ExtractRequestUser(c)
+
+	if reqUser == nil {
+		c.AbortWithStatus(http.StatusUnauthorized)
+
+		return "", nil
+	}
+
+	userStringRole, foundPermission, err := CheckUserFilePermission(c,
+		fileClient,
+		permissionClient,
+		reqUser.ID,
+		fileID,
+		role)
+
+	if err != nil {
+		httpStatusCode := gwruntime.HTTPStatusFromCode(status.Code(err))
+		loggermiddleware.LogError(r.logger, c.AbortWithError(httpStatusCode, err))
+
+		return "", nil
+	}
+
+	if userStringRole == "" {
+		c.AbortWithStatus(http.StatusUnauthorized)
+	}
+
+	return userStringRole, foundPermission
+}
+
+
+// CheckUserFilePermission checks if userID is permitted to fileID with the wanted role.
+// The function returns the role name if the user is permitted to the file,
+// the permission if the user was shared, and non-nil err if any encountered.
+// If no permitted then role found role would be "".
+// If user was shared then permission would be non-nil.
+func CheckUserFilePermission(c context.Context,
+	fileClient fpb.FileServiceClient,
+	permissionClient ppb.PermissionClient,
+	userID string,
+	fileID string,
+	role ppb.Role) (string, *ppb.PermissionObject, error) {
+	if userID == "" {
+		return "", nil, fmt.Errorf("userID is required")
+	}
+
+	// Everyone is permitted to their root, since all actions on root are authenticated,
+	// and it's impossible to create a permission for root (aka sharing a user's whole drive).
+	if fileID == "" {
+		return OwnerRole, nil, nil
+	}
+
+	// Get the file's metadata.
+	file, err := fileClient.GetFileByID(c, &fpb.GetByFileByIDRequest{Id: fileID})
+	if err != nil {
+		return "", nil, err
+	}
+
+	// Check if the owner of the current file is userID, if so then he's permitted.
+	if file.GetOwnerID() == userID {
+		return OwnerRole, nil, nil
+	}
+
+	// Go up the hierarchy searching for a permission for userID to fileID with role.
+	// Fetch fileID's parents, each at a time, and check permission to each parent.
+	// If reached a parent that userID isn't permitted to then return with error,
+	// If reached a parent that userID is permitted to then return true with nil error.
+	// If any error encountered then return false and the encountered error.
+	currentFile := fileID
+	for {
+		// If reached the root and didn't find a permission then userID is not permitted to fileID.
+		if currentFile == "" {
+			return "", nil, nil
+		}
+
+		// Check if the user has an existing permission and is permitted to currentFile with the wanted role.
+		isPermitted, err := permissionClient.IsPermitted(c,
+			&ppb.IsPermittedRequest{FileID: currentFile, UserID: userID, Role: role})
+
+		// If an error occurred which is NOT grpc's NotFound error which
+		// indicates that the permission doesn't not exist.
+		if err != nil && status.Code(err) != codes.NotFound {
+			return "", nil, err
+		}
+
+		// If no error received and user isn't permitted.
+		if !isPermitted.GetPermitted() && err == nil {
+			return "", nil, nil
+		}
+
+		// If userID is permitted with the wanted role then return the role that the user has for the file.
+		if isPermitted.GetPermitted() {
+			permission, err := permissionClient.GetPermission(
+				c,
+				&ppb.GetPermissionRequest{
+					FileID: currentFile,
+					UserID: userID,
+				},
+			)
+
+			if err != nil {
+				return "", nil, err
+			}
+
+			return permission.GetRole().String(), permission, nil
+		}
+
+		// Get the current file's metadata.
+		file, err := fileClient.GetFileByID(c, &fpb.GetByFileByIDRequest{Id: currentFile})
+		if err != nil {
+			return "", nil, err
+		}
+
+		// Repeat for the file's parent.
+		currentFile = file.GetParent()
+	}
 }