+++ CHANGELOG.md
@@ -7,11 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- hotfix([262](https://github.com/meateam/api-gateway/pull/262)): change HandleUserFilePermission and CheckUserFilePermission functions to generic in utils.
+
 ## [v5.1.0] - 2021-12-7
 
 ### Added 
 
-- minor([246](https://github.com/meateam/api-gateway/pull/246)) : add fav service
+- minor([246](https://github.com/meateam/api-gateway/pull/246)): add fav service
 
 ## [v5.0.2] - 2021-08-26